Rename *_defines.h => *_constants.h
