+++ src/solver/optimisation/opt_structure_probleme_a_resoudre.h
@@ -29,7 +29,7 @@
 
 #include <vector>
 #include <string>
-#include "opt_defines.h"
+#include "opt_constants.h"
 
 /*--------------------------------------------------------------------------------------*/
