+++ src/solver/simulation/sim_structure_probleme_economique.h
@@ -32,7 +32,7 @@
 #include "../utils/optimization_statistics.h"
 #include "../../libs/antares/study/fwd.h"
 #include "../../libs/antares/study/study.h"
-#include "sim_defines.h"
+#include "sim_constants.h"
 #include <memory>
 #include <yuni/core/math.h>
